hwmon: ntc_thermistor: correct set_emul_temp prototype
This patch corrects the ntc_set_emul_temp function prototype
to address below build warning.

drivers/hwmon/ntc_thermistor.c:549:19: warning: initialization from incompatible pointer type
  .set_emul_temp = ntc_set_emul_temp,
                   ^
drivers/hwmon/ntc_thermistor.c:549:19: warning: (near initialization for 'ntc_of_thermal_ops.set_emul_temp')

+++ drivers/hwmon/ntc_thermistor.c
@@ -493,7 +493,7 @@ static int ntc_read_temp(void *dev, int *temp)
 	return 0;
 }
 
-static int ntc_set_emul_temp(void *dev, unsigned long temp)
+static int ntc_set_emul_temp(void *dev, int temp)
 {
 	struct ntc_data *data = dev_get_drvdata(dev);
